applying modification recommended by bito

+++ notes_impl/dn_impl.go
@@ -1,12 +1,13 @@
 package notes_impl
 
 import (
-	"errors"
+	"net/http"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 	"github.com/rocha7778/dynamo-db/db"
+	"github.com/rocha7778/dynamo-db/validations"
 	"github.com/rocha7778/dynamo-db/variables"
 )
 
@@ -16,26 +17,24 @@ type DefaultNoteDeleteService struct {
 type DeleteServiceRepository struct{}
 
 // DeleteNote deletes a note
-func (NoteService *DefaultNoteDeleteService) DeleteNote(noteID string) (events.APIGatewayProxyResponse, error) {
-	if noteID == "" {
-		return events.APIGatewayProxyResponse{StatusCode: 400, Body: "Note ID is required in path parameters"}, errors.New("Note ID is required in path parameters")
+func (NoteService *DefaultNoteDeleteService) DeleteNote(noteID string) events.APIGatewayProxyResponse {
+	if noteID == "" || !validations.IsValidNoteID(noteID) {
+		return handleError("Note ID is required in path parameters", http.StatusBadRequest)
 	}
 	err := NoteService.Repo.DeleteItem(noteID)
 
 	if err != nil {
-		return events.APIGatewayProxyResponse{StatusCode: 500, Body: "Error deleting note from DynamoDB"}, err
+		return handleError("Error deleting note from DynamoDB", http.StatusInternalServerError)
 	}
 
-	return events.APIGatewayProxyResponse{StatusCode: 204, Body: "Note deleted successfully"}, nil
+	return events.APIGatewayProxyResponse{StatusCode: 204, Body: "Note deleted successfully"}
 }
 
 func (*DeleteServiceRepository) DeleteItem(noteId string) error {
-
-	// Delete the item from DynamoDB
 	_, err := db.DBClient().DeleteItem(&dynamodb.DeleteItemInput{
 		TableName: aws.String(variables.TableName),
 		Key: map[string]*dynamodb.AttributeValue{
-			"id": {S: aws.String(noteId)},
+			"id": {S: &noteId},
 		},
 	})
 	return err
